+++ app/config/config.py
@@ -3,15 +3,15 @@ import json
 import logging
 import os
 
+import botocore.session
+from aws_secretsmanager_caching import SecretCache, SecretCacheConfig
 from dotenv import load_dotenv
+
+from utils.logging import setup_logging
+from utils.slack_alert import init_slack
 
 # Load environment variables from .env file
 load_dotenv()
-
-import botocore.session
-from aws_secretsmanager_caching import SecretCache, SecretCacheConfig
-
-from utils.logging import setup_logging
 
 logger = logging.getLogger(__name__)
 
@@ -60,6 +60,9 @@ class Config:
         # this part can be load from env or aws secrets manager
         self.db["auto_migrate"] = self.load("DB_AUTO_MIGRATE", "true") == "true"
         self.debug = self.load("DEBUG") == "true"
+        self.debug_resp = (
+            self.load("DEBUG_RESP", "false") == "true"
+        )  # Agent response with thought steps and time cost
         self.cdp_api_key_name = self.load("CDP_API_KEY_NAME")
         self.cdp_api_key_private_key = self.load("CDP_API_KEY_PRIVATE_KEY")
         self.openai_api_key = self.load("OPENAI_API_KEY")
@@ -69,14 +72,23 @@ class Config:
         self.tg_server_host = self.load("TG_SERVER_HOST", "127.0.0.1")
         self.tg_server_port = self.load("TG_SERVER_PORT", "8081")
         self.tg_new_agent_poll_interval = self.load("TG_NEW_AGENT_POLL_INTERVAL", "60")
-
+        self.twitter_endpoint_interval = int(
+            self.load("TWITTER_ENDPOINT_INTERVAL", "15")
+        )  # in minutes
+        self.slack_alert_token = self.load(
+            "SLACK_ALERT_TOKEN"
+        )  # For alert purposes only
+        self.slack_alert_channel = self.load("SLACK_ALERT_CHANNEL")
         # Now we know the env, set up logging
         setup_logging(self.env, self.debug)
         logger.info("config loaded")
+        # If the slack alert token exists, init it
+        if self.slack_alert_token and self.slack_alert_channel:
+            init_slack(self.slack_alert_token, self.slack_alert_channel)
 
     def load(self, key, default=None):
         """Load a secret from the secrets map or env"""
         return self.secrets.get(key, os.getenv(key, default))
 
 
-config = Config()
+config: Config = Config()