+++ lint.sh
@@ -1,3 +1,2 @@
-black .
-isort .
-bandit -r app/ skills/ skill_sets/ utils/ -ll
+ruff format
+ruff check --fix