+++ CHANGELOG.md
@@ -1,5 +1,39 @@
 # Changelog
 
+## 2024-12-28
+
+### New Features
+- Twitter Entrypoint
+- Admin cron for quota clear
+- Admin API get all agents
+
+### Improvements
+- Change lint tools to ruff
+- Improve CI
+- Improve twitter skills
+
+### Bug Fixes
+- Fix bug in db base code
+
+## 2024-12-27
+
+### New Features
+- Twitter Skills
+    - Get Mentions
+    - Get Timeline
+    - Post Tweet
+    - Reply Tweet
+
+### Improvements
+- CI/CD refactoring for better security
+
+## 2024-12-26
+
+### Improvements
+- Change default plan to "self-hosted" from "free", new agent now has 9999 message limit for testing
+- Add a flag "DEBUG_RESP", when set to true, the Agent will respond with thought processes and time costs
+- Better DB session management
+
 ## 2024-12-25
 
 ### Improvements